+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 * 22.0.0
   - Strongly convex functionality in TotalVariation and FGP_TV Functions.
+  - Refactored KullbackLeibler function class. Fix bug on gradient method for SIRF objects.
+  - Numba added as a CIL requirement.
   - Simplify initialisation of `CentreOfRotation.ImageSharpness` with new parameter `backend`
   - Added ISTA algorithm. Improve inheritance of proximal gradient algorithms.
   - Updated interface to `plugins.tigre`/`plugins.astra` `FBP` and `ProjectionOperator` classes